package updates + add vue router
