+++ package.json
@@ -14,14 +14,15 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.22",
+    "@headlessui/vue": "^1.7.23",
     "@heroicons/vue": "^2.1.5",
-    "@vueuse/core": "^10.11.0",
+    "@vueuse/core": "^11.2.0",
     "lz-string": "^1.5.0",
-    "pinia": "^2.1.7",
-    "uuid": "^10.0.0",
-    "vue": "^3.4.29",
-    "vue-confetti-explosion": "^1.0.2"
+    "pinia": "^2.2.5",
+    "uuid": "^11.0.3",
+    "vue": "^3.5.13",
+    "vue-confetti-explosion": "^1.0.2",
+    "vue-router": "4"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.8.0",
@@ -37,17 +38,17 @@
     "autoprefixer": "^10.4.20",
     "eslint": "^8.57.0",
     "eslint-plugin-vue": "^9.23.0",
-    "jsdom": "^24.1.0",
+    "jsdom": "^25.0.1",
     "npm-run-all2": "^6.2.0",
     "postcss": "^8.4.40",
     "postcss-import": "^16.1.0",
     "prettier": "^3.2.5",
     "sass": "^1.77.8",
     "sass-loader": "^16.0.0",
-    "tailwindcss": "^3.4.7",
-    "typescript": "~5.4.0",
-    "vite": "^5.3.1",
-    "vitest": "^1.6.0",
-    "vue-tsc": "^2.0.21"
+    "tailwindcss": "^3.4.15",
+    "typescript": "~5.6.3",
+    "vite": "^5.4.11",
+    "vitest": "^2.1.5",
+    "vue-tsc": "^2.1.10"
   }
 }
